+++ site/app/templates/forum/ForumBar.twig
@@ -66,8 +66,7 @@
 				{% set moreItems = moreItems|merge([{
 						'id': 'mark-unread',
 						'displayText': 'Unread Thread',
-						'title': 'Mark Thread Unread',
-						'dividerAfter': true
+						'title': 'Mark Thread Unread'
 					}]) %}
 			{% endif %}
 			{% if more_data is defined %}
@@ -77,8 +76,7 @@
 							'id': more.id,
 							'displayText': more.display_text,
 							'title': more.title,
-							'link': more.link,
-							'optionalClass': more.optional_class[1] ?? ''
+							'link': more.link
 						}]) %}
 					{% endif %}
 				{% endfor %}
@@ -88,8 +86,7 @@
 					'id': 'edit-categories',
 					'displayText': 'Edit Categories',
 					'title': 'Edit Forum Categories',
-					'link': manage_categories_url,
-					'dividerBefore': true
+					'link': manage_categories_url
 				}]) %}
 			{% endif %}
 			{% if thread_exists and (is_full_threads_page is not defined or not is_full_threads_page) %}
@@ -97,29 +94,32 @@
 						'id': 'toggle-attachments',
 						'displayText': 'Show Attachments',
 						'title': 'Click to toggle all post attachments',
-						'badgeText': total_attachments is defined ? ''~total_attachments : '',
-						'dividerBefore': true
+						'badgeText': total_attachments is defined ? ''~total_attachments : ''
 					}]) %}
+				{% set currentDisplayOption = display_option is defined ? display_option : '' %}
+				{% set moreItems = moreItems|merge([
+					{ 'id': 'display-tree', 'displayText': 'Hierarchical Post Order', 'title': 'Sort posts by reply hierarchy' },
+					{ 'id': 'display-time', 'displayText': 'Chronological Post Order (Asc)', 'title': 'Sort posts by ascending chronological order' },
+					{ 'id': 'display-reverse-time', 'displayText': 'Chronological Post Order (Desc)', 'title': 'Sort posts by descending chronological order' }
+				]) %}
+				{% if user_group <= core.getUser().accessGrading() %}
+					{% set moreItems = moreItems|merge([
+						{ 'id': 'display-alpha', 'displayText': 'Alphabetical Post Order', 'title': "Sort posts by author's last name" },
+						{ 'id': 'display-alpha_by_registration', 'displayText': 'Alpha by Registration Post Order', 'title': 'Sort posts by author registration section then last name' },
+						{ 'id': 'display-alpha_by_rotating', 'displayText': 'Alpha by Rotating Post Order', 'title': 'Sort posts by author rotating section then last name' }
+					]) %}
+				{% endif %}
 			{% endif %}
-			{% set moreDropdownArgs = {
-				'items': moreItems,
-				'currentDisplayOption': display_option is defined ? display_option : '',
-				'threadExists': thread_exists is defined and thread_exists,
-				'isFullThreadsPage': is_full_threads_page is defined and is_full_threads_page,
-			} %}
 			{% include "Vue.twig" with {
 				'type': 'component',
-				'name': 'forum/MoreDropdown',
-				'args': moreDropdownArgs,
-				'class': 'more-dropdown',
+				'name': 'DropdownButton',
+				'args': {
+					'label': 'More',
+					'items': moreItems
+				},
 				'events': {
-					'display-option-change': 'changeDisplayOptions',
-					'toggle-merged': 'forumToggleMerged',
-					'toggle-deleted': 'forumToggleDeleted',
-					'navigate-stats': 'forumNavigateStats',
-					'navigate': '(url) => { window.location.href = url; }',
-					'toggle-attachments': 'loadAllInlineImages',
-					'item-click': 'forumHandleMoreItemClick',
+					'select': 'forumHandleMoreDropdownSelect',
+					'navigate': '(url) => { window.location.href = url; }'
 				}
 			} %}
 			{% endif %}
