+++ site/public/js/forum.js
@@ -2740,27 +2740,26 @@ function highlightAndScrollToCurrentThread() {
     }
 }
 
-// Wrappers for Vue component event handlers that read state from DOM/cookies.
-function forumToggleMerged() {
-    const course = document.body.dataset.courseUrl.split('/').filter(Boolean).pop();
-    const current = Cookies.get(`${course}_show_merged_thread`);
-    alterShowMergeThreadStatus(current === '1' ? 0 : 1, course);
-}
-function forumToggleDeleted() {
-    const current = Cookies.get('show_deleted');
-    alterShowDeletedStatus(current === '1' ? 0 : 1);
-}
-function forumMarkUnread() {
-    markThreadUnread($('#current-thread').val());
-}
-function forumHandleMoreItemClick(itemId) {
+function forumHandleMoreDropdownSelect(itemId) {
     if (itemId === 'mark-unread') {
-        forumMarkUnread();
+        markThreadUnread($('#current-thread').val());
     }
     else if (itemId === 'toggle-attachments') {
         loadAllInlineImages();
     }
-}
-function forumNavigateStats() {
-    window.location.href = buildCourseUrl(['forum', 'stats']);
+    else if (itemId === 'merge_thread') {
+        const course = document.body.dataset.courseUrl.split('/').filter(Boolean).pop();
+        const current = Cookies.get(`${course}_show_merged_thread`);
+        alterShowMergeThreadStatus(current === '1' ? 0 : 1, course);
+    }
+    else if (itemId === 'delete') {
+        const current = Cookies.get('show_deleted');
+        alterShowDeletedStatus(current === '1' ? 0 : 1);
+    }
+    else if (itemId === 'forum_stats') {
+        window.location.href = buildCourseUrl(['forum', 'stats']);
+    }
+    else if (itemId.startsWith('display-')) {
+        changeDisplayOptions(itemId.substring('display-'.length));
+    }
 }
